core/thread.c: store actual stack size
Before `thread->stack_size` also accounted for overhead such as the
`thread_t` structure, the thread local storage, and any (potential)
architecture specific per-thread housekeeping data structures allocated
inside the thread stack.

This changes the accounting to only account for the stack that is
left after the thread is freshly created and before any code of the
thread is computed.