sys/net/nanocoap: fix UB when building hdr
Some callers of `coap_build_hdr()` are doing so with the target buffer
already containing the token passed as argument: They reuse the buffer
that held the request to assemble the response in. We cannot rely
on `memcpy()` in this case to copy the token into the target buffer, as
source and destination would (fully) overlap.

This makes reusing the request buffer for the response a special case
and just not copy the token in this case. This is possibly more
efficient than just using `memmove()`, as `memmove()` will also handle
cases where `src` and `dest` are partially overlapping (e.g. have an
offset of 1).